fix: Cleanup webroot handling
When placing the webroot on an alternate drive (e.g. `W:`) then trailing
colon will be interpretted as a yaml mapping and results in an error.

Add a trailing `\` in-case the webroot was specified as only a drive
letter.

Do not attempt to change perms on the webroot if the webroot points at a
drive w/out a directory.  E.g. `W:\`